move contracts from lib/api to app/contracts
