+++ app/contracts/work_packages/create_contract.rb
@@ -27,24 +27,20 @@
 # See doc/COPYRIGHT.rdoc for more details.
 #++
 
-module API
-  module V3
-    module WorkPackages
-      class CreateContract < BaseContract
-        # These attributes need to be set during creation and cannot be modified via representer.
-        # Hence making them writable here is unproblematic.
-        attribute :project_id
-        attribute :author_id
+module WorkPackages
+  class CreateContract < BaseContract
+    # These attributes need to be set during creation and cannot be modified via representer.
+    # Hence making them writable here is unproblematic.
+    attribute :project_id
+    attribute :author_id
 
-        validate :user_allowed_to_add
+    validate :user_allowed_to_add
 
-        private
+    private
 
-        def user_allowed_to_add
-          unless @user.allowed_to?(:add_work_packages, model.project)
-            errors.add :base, :error_unauthorized
-          end
-        end
+    def user_allowed_to_add
+      unless @user.allowed_to?(:add_work_packages, model.project)
+        errors.add :base, :error_unauthorized
       end
     end
   end
