+++ lib/api/v3/work_packages/create_form_api.rb
@@ -27,7 +27,7 @@
 #++
 
 require 'api/v3/work_packages/work_packages_shared_helpers'
-require 'api/v3/work_packages/create_contract'
+require 'work_packages/create_contract'
 
 module API
   module V3
@@ -37,7 +37,7 @@ module API
           helpers ::API::V3::WorkPackages::WorkPackagesSharedHelpers
 
           post do
-            create_contract = ::API::V3::WorkPackages::CreateContract
+            create_contract = ::WorkPackages::CreateContract
             create_work_package_form(create_service.create,
                                      contract_class: create_contract,
                                      form_class: CreateFormRepresenter)
