+++ spec/lib/api/v3/work_packages/work_packages_shared_helpers_spec.rb
@@ -66,7 +66,7 @@ describe ::API::V3::WorkPackages::WorkPackagesSharedHelpers do
   describe '#create_work_package_form' do
     subject do
       helper.create_work_package_form(work_package,
-                                      contract_class: ::API::V3::WorkPackages::CreateContract,
+                                      contract_class: ::WorkPackages::CreateContract,
                                       form_class: ::API::V3::WorkPackages::CreateFormRepresenter)
     end
